requirements.txt
jieba
sqlalchemy
fuzzywuzzy
python-Levenshtein
# ext
python-forecastio
geocoder